The 2289-sample pollen data-set, in the form used with BRT.
This includes both climate and pollen data for the 2289 samples.

ID1: Sample number.
tjul: July mean T.
clim2: Values for the second climatic variable to be used - this is full of zeroes for now.
v1...v56: Values for the 56 pollen types. These are the same taxa as in the original data, and in the same order. I removed a few pollen types which had only zero abundances in this 2289-sample subset of the original Whitmore et al. data-set.